add generate init function

+++ generate/generate.php
@@ -92,18 +92,152 @@ class Generate{
         }
     }
     function initApp(){
+        $appDir = APP_PATH.DIRECTORY_SEPARATOR.'app'.DIRECTORY_SEPARATOR;
         /*
-        mkdir(APP_PATH);
+           make app dir
+         */
+
+        mkdir($appDir);
+
+        /*
+           make views dir
+         */
+
         mkdir(VIEWS_PATH);
+
+
+        /*
+           make controllers dir
+         */
+
+        mkdir(CONTROLLERS_PATH);
+
+
+        /*
+           make models dir
+         */
+
+        mkdir(MODELS_PATH);
+
+
+        /*
+           make layouts dir
+         */
         mkdir(LAYOUTS_PATH);
-        mkdir(ASSETS_PATH);
-        */
+
+
+        /*
+           make assets dir
+         */
+
+        mkdir($appDir.ASSETS_PATH);
+
+        /*
+           make javascripts dir
+         */
+
+        mkdir($appDir.JAVASCRIPTS_PATH);
+
+        /*
+           make images dir
+         */
+
+        mkdir($appDir.IMAGES_PATH);
+
+        /*
+           make stylesheets dir
+         */
+
+        mkdir($appDir.STYLESEETS_PATH);
+
+        /*
+           make config dir and config files
+         */
+
+        $this -> createConfig();
+
+        /*
+           make public dir
+         */
+
+        $this -> createPublic();
+
+        /*
+           make app index files 
+         */
+
+        $this -> createIndex();
+
+        /*
+           make data files 
+         */
+
+        $this -> createData();
+
+        /*
+           make app files 
+         */
+
+        $this -> scaffold('home');
+    }
+    function createData(){
+    }
+    function createIndex(){
+        $index = APP_PATH.DIRECTORY_SEPARATOR.'index.php';
+        $tpl = $this -> getTpl('index');
+        $return = file_put_contents($index,$tpl);
+        if($return){
+            $this -> throwMsg('success','index file created',$index);
+        }else{
+            $this -> throwMsg('error','Error','Can\'t create index file');
+        }
+    }
+    function createPublic(){
+        mkdir(PUBLIC_PATH);
+        $_404File = PUBLIC_PATH.'404.html';
+        $_404tpl = $this->getTpl('404');
+        $_404Return = file_put_contents($_404File,$_404tpl);
+        if($_404Return){
+            $this -> throwMsg('success','404 file created',$_404File);
+        }else{
+            $this -> throwMsg('error','Error','Can\'t create 404 file');
+        }
+    }
+    function createConfig(){
+        mkdir(CONFIG_PATH);
+        $configFile = CONFIG_PATH.'config.php';
+        $dbFile = CONFIG_PATH.'database.php';
+        $configtpl = $this->getTpl('config');
+        $dbtpl = $this->getTpl('database');
+        $configReturn = file_put_contents($configFile,$configtpl);
+        if($configReturn){
+            $this -> throwMsg('success','Config file created',$configFile);
+        }else{
+            $this -> throwMsg('error','Error','Can\'t create config file');
+        }
+        $dbReturn = file_put_contents($dbFile,$dbtpl);
+        if($dbReturn){
+            $this -> throwMsg('success','Database config file created',$dbFile);
+        }else{
+            $this -> throwMsg('error','Error','Can\'t create database config file');
+        }
     }
     function createController($name){
-        $tpl = $this -> get_tpl('controller');
+        $appliction = CONTROLLERS_PATH.'application_controller.php';
+        if(!file_exists($appliction)){
+            $app = $this -> getTpl('application_controller');
+            $app = str_replace('<{name}>',$name,$app);
+            $appReturn = file_put_contents($appliction,$app);
+            if($appReturn){
+                $this -> throwMsg('success','appliction controller created',$appliction);
+            }else{
+                $this -> throwMsg('error','Error','Can\'t create appliction controller file');
+            }
+        }
+        $tpl = $this -> getTpl('controller');
         $tpl = str_replace('<{name}>',ucfirst($name),$tpl);
         $cltName = $name.'_controller.php';
-        $cltFile = APP_PATH.DIRECTORY_SEPARATOR.'app'.DIRECTORY_SEPARATOR.'controllers'.DIRECTORY_SEPARATOR.$cltName;
+        $cltFile = CONTROLLERS_PATH.$cltName;
         if(file_exists($cltFile)){
             $this -> throwMsg('error','Controller exists!',"controller file: $cltName exists!");
         }else{
@@ -117,10 +251,10 @@ class Generate{
         }
     }
     function createModel($name){
-        $tpl = $this -> get_tpl('model');
+        $tpl = $this -> getTpl('model');
         $tpl = str_replace('<{name}>',ucfirst($name),$tpl);
         $mdName = $name.'.php';
-        $mdFile = APP_PATH.DIRECTORY_SEPARATOR.'app'.DIRECTORY_SEPARATOR.'models'.DIRECTORY_SEPARATOR.$mdName;
+        $mdFile = MODELS_PATH.$mdName;
         if(file_exists($mdFile)){
             $this -> throwMsg('error','Model exists!',"Model file: $mdName exists!");
         }else{
@@ -133,12 +267,22 @@ class Generate{
         }
     }
     function createView($controller,$name = 'index'){
-        $tpl = $this -> get_tpl('view');
+        $appliction = LAYOUTS_PATH.'application.html';
+        if(!file_exists($appliction)){
+            $app = $this -> getTpl('application_view');
+            $appReturn = file_put_contents($appliction,$app);
+            if($appReturn){
+                $this -> throwMsg('success','appliction view created',$appliction);
+            }else{
+                $this -> throwMsg('error','Error','Can\'t create appliction view file');
+            }
+        }
+        $tpl = $this -> getTpl('view');
         $tpl = str_replace('<{name}>',$name,$tpl);
         $tpl = str_replace('<{controller}>',$controller,$tpl);
         $viName = $controller.DIRECTORY_SEPARATOR.$name.'.html';
-        $viFile = APP_PATH.DIRECTORY_SEPARATOR.'app'.DIRECTORY_SEPARATOR.'views'.DIRECTORY_SEPARATOR.$viName;
-        $viDir = APP_PATH.DIRECTORY_SEPARATOR.'app'.DIRECTORY_SEPARATOR.'views'.DIRECTORY_SEPARATOR.$controller;
+        $viFile = VIEWS_PATH.$viName;
+        $viDir = VIEWS_PATH.$controller;
         if(file_exists($viFile)){
             $this -> throwMsg('error','View exists!',"view file: $viName exists!");
         }else{
@@ -154,28 +298,39 @@ class Generate{
         }
     }
     function createRoute($name){
-        $routeFile = APP_PATH.DIRECTORY_SEPARATOR.'config'.DIRECTORY_SEPARATOR.'routes.php';
-        $route = require_once($routeFile);
-        if(isset($route[$name])){
-            $this -> throwMsg('error','Route exists!','route: '.$name.' exists');
-        }else{
-            $oldRoute = file_get_contents($routeFile);
-            $routeMatch = '/(<?[\s\S]*)(?:\)\;)/';
-            preg_match($routeMatch,$oldRoute,$matchs);
-            $check = preg_match('/(\,|,\s*)$/',$matchs[1]);
-            if(!$check){
-                $endRoute = $matchs[1].',';
-            }else{
-                $endRoute = $matchs[1];
-            }
-            $newRoute = "\n    '".$name."' => array(\n        'index' => array(),\n    ),\n);\n?>";
-            $endRoute .= $newRoute;
-            $return = file_put_contents($routeFile,$endRoute);
+        $routeFile = CONFIG_PATH.'routes.php';
+        if(!file_exists($routeFile)){
+            $tpl = $this->getTpl('route');
+            $tpl = str_replace('<{name}>',$name,$tpl);
+            $return = file_put_contents($routeFile,$tpl);
             if($return){
-                $this -> throwMsg('success','Route created',$name."=> array(\n'index' => array()\n)");
+                $this -> throwMsg('success','Route file created',$routeFile);
             }else{
-                $this -> throwMsg('error','Error','Can\'t create route');
+                $this -> throwMsg('error','Error','Can\'t create route file');
             }
+        }else{
+            $route = require_once($routeFile);
+            if(isset($route[$name])){
+                $this -> throwMsg('error','Route exists!','route: '.$name.' exists');
+            }else{
+                $oldRoute = file_get_contents($routeFile);
+                $routeMatch = '/(<?[\s\S]*)(?:\)\;)/';
+                preg_match($routeMatch,$oldRoute,$matchs);
+                $check = preg_match('/(\,|,\s*)$/',$matchs[1]);
+                if(!$check){
+                    $endRoute = $matchs[1].',';
+                }else{
+                    $endRoute = $matchs[1];
+                }
+                $newRoute = "\n    '".$name."' => array(\n        'index' => array(),\n    ),\n);\n?>";
+                $endRoute .= $newRoute;
+                $return = file_put_contents($routeFile,$endRoute);
+                if($return){
+                    $this -> throwMsg('success','Route created',$name."=> array(\n'index' => array()\n)");
+                }else{
+                    $this -> throwMsg('error','Error','Can\'t create route');
+                }
+            };
         }
     }
     function scaffold($name){
@@ -186,7 +341,7 @@ class Generate{
     function createHelp($name){
         //todo
     }
-    function get_tpl($tpl){
+    function getTpl($tpl){
         return file_get_contents(__DIR__.DIRECTORY_SEPARATOR.'templates'.DIRECTORY_SEPARATOR.$tpl.'.tpl');
     }
     function throwMsg($type,$title,$msg){
